Split into one fixture and test per cross-database macro
