+++ tests/functional/cross_db_utils/fixture_cross_db_macro.py
@@ -1,0 +1,6 @@
+macros__test_assert_equal_sql = """
+{% test assert_equal(model, actual, expected) %}
+select * from {{ model }} where {{ actual }} != {{ expected }}
+
+{% endtest %}
+"""
